Crate keycodes_ascii

Structs

KeyEvent

The KeyEvent that should be delivered to applications upon a keyboard action.

KeyboardModifiers

The set of modifier keys that can be held down while other keys are pressed.

Enums

KeyAction

Whether a keyboard event was a key press or a key released.

Keycode

Constants

KEY_RELEASED_OFFSET

The offset that a keyboard adds to the scancode to indicate that the key was released rather than pressed. So if a scancode of 1 means a key foo was pressed, a scancode of 129 (1 + 128) means that key foo was released.

Functions

scancode_to_ascii

convenience function for obtaining the ascii value for a raw scancode under the given modifiers