Enum keycodes_ascii::Keycode

#[repr(u8)]
pub enum Keycode {

    OverflowError,
    Escape,
    Num1,
    Num2,
    Num3,
    Num4,
    Num5,
    Num6,
    Num7,
    Num8,
    Num9,
    Num0,
    Minus,
    Equals,
    Backspace,
    Tab,
    Q,
    W,
    E,
    R,
    T,
    Y,
    U,
    I,
    O,
    P,
    LeftBracket,
    RightBracket,
    Enter,
    Control,
    A,
    S,
    D,
    F,
    G,
    H,
    J,
    K,
    L,
    Semicolon,
    Quote,
    Backtick,
    LeftShift,
    Backslash,
    Z,
    X,
    C,
    V,
    B,
    N,
    M,
    Comma,
    Period,
    Slash,
    RightShift,
    PadMultiply,
    Alt,
    Space,
    CapsLock,
    F1,
    F2,
    F3,
    F4,
    F5,
    F6,
    F7,
    F8,
    F9,
    F10,
    NumLock,
    ScrollLock,
    Home,
    Up,
    PageUp,
    PadMinus,
    Left,
    Pad5,
    Right,
    PadPlus,
    End,
    Down,
    PageDown,
    Insert,
    Delete,
    Unknown1,
    Unknown2,
    NonUsBackslash,
    F11,
    F12,
    Pause,
    Unknown3,
    SuperKeyLeft,
    SuperKeyRight,
    Menu,
    ControlReleased,
    AltReleased,
    LeftShiftReleased,
    RightShiftReleased,
    SuperKeyLeftReleased,
    SuperKeyRightReleased,

}

Variants

OverflowError

Escape

Num1

Num2

Num3

Num4

Num5

Num6

Num7

Num8

Num9

Num0

Minus

Equals

Backspace

Tab

Q

W

E

R

T

Y

U

I

O

P

LeftBracket

RightBracket

Enter

Control

A

S

D

F

G

H

J

K

L

Semicolon

Quote

Backtick

LeftShift

Backslash

Z

X

C

V

B

N

M

Comma

Period

Slash

RightShift

PadMultiply

Alt

Space

CapsLock

F1

F2

F3

F4

F5

F6

F7

F8

F9

F10

NumLock

ScrollLock

Home

Up

PageUp

PadMinus

Left

Pad5

Right

PadPlus

End

Down

PageDown

Insert

Delete

Unknown1

Unknown2

NonUsBackslash

F11

F12

Pause

Unknown3

SuperKeyLeft

SuperKeyRight

Menu

ControlReleased

AltReleased

LeftShiftReleased

RightShiftReleased

SuperKeyLeftReleased

SuperKeyRightReleased

Implementations

impl Keycode

pub fn to_ascii(&self, modifiers: KeyboardModifiers) -> Option<char>

Obtains the ascii value for a keycode under the given modifiers

pub fn is_letter(&self) -> bool

returns true if this keycode was a letter from A-Z